pyproject.toml
[tool.poetry]
name = "chat-house"
version = "0.1.0"
description = ""
authors = ["Stefan Jacobs <stefan.jacobs@deutschebahn.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^21.6"
requests = "^2.32.3"
openai = "^1.50.1"
pydantic-core = "^2.23.3"
python-dotenv = "^1.0.1"
dwdwfsapi = "^1.1.0"
aiosqlite = "^0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
